moved RemoveUnreferencedVerts
